Move adapters to their own module
