+++ amiadapters/adapters/base.py
